Handle buffer in Packet class

+++ ServerCore/Session.cs
@@ -9,7 +9,7 @@ namespace ServerCore
 {
     public abstract class PacketSession : Session
     {
-        public static readonly int HeaderSize = 2;
+        public static readonly ushort HeaderSize = 2;
 
         // [size(2)][packetId(2)][.....][size(2)][packetId(2)][.....]
         public sealed override int OnRecv(ArraySegment<byte> buffer)
